% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{bsa.exact.wrapper}
\alias{bsa.exact.wrapper}
\title{This function wraps around bsa.exactchgpt.}
\usage{
bsa.exact.wrapper(X, Y, kmax, delta)
}
\arguments{
\item{X}{an n*p data matrix, n=observation, p=dimension}

\item{Y}{an n dimensional data vector, n=observation}

\item{kmax}{the maximum change point number.}

\item{delta}{the minimal interval length specified by users.}
}
\value{
A list including: the estimated change point number, the estimated
change point locations.
}
\description{
This function wraps around bsa.exactchgpt.
}
