% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{bsachgpt}
\alias{bsachgpt}
\title{This function provides change point estimation by Binary Segmentation Algorithm
for the high dimensional generalized linear regression model.}
\usage{
bsachgpt(X, Y, kmax, c1, m1, delta)
}
\arguments{
\item{X}{an n*p data matrix, n=observation, p=dimension}

\item{Y}{an n dimensional data vector, n=observation}

\item{kmax}{the maximum change point number.}

\item{c1}{a constant specified by users.}

\item{m1}{the sample size of the minimal interval.}

\item{delta}{the minimal interval length specified by users.}
}
\value{
A list including: the estimated change point number, the estimated
change point locations.
}
\description{
This function provides change point estimation by Binary Segmentation Algorithm
for the high dimensional generalized linear regression model.
}
