% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Compare with GLM}
\usage{
plot_comparison(fit_ours, fit_glm)
}
\arguments{
\item{fit_ours}{Our logistic_fit object}

\item{fit_glm}{A glm object}
}
\value{
A ggplot object
}
\description{
Compare coefficients with glm().
}
