% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fitted}
\alias{plot_fitted}
\title{Plot Fitted Values}
\usage{
plot_fitted(fit)
}
\arguments{
\item{fit}{A logistic_fit object}
}
\value{
A ggplot object
}
\description{
Plot fitted probabilities vs observed values.
}
\examples{
set.seed(123)
X <- matrix(rnorm(100), ncol = 2)
y <- rbinom(50, 1, 0.5)
fit <- logistic_fit(X, y)
plot_fitted(fit)

}
