% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.logistic_fit}
\alias{predict.logistic_fit}
\title{Predict from Logistic Fit}
\usage{
\method{predict}{logistic_fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A logistic_fit object}

\item{newdata}{New data matrix (optional)}

\item{...}{Not used}
}
\value{
Vector of predicted probabilities
}
\description{
Predict probabilities from fitted model.
}
\examples{
set.seed(123)
X <- matrix(rnorm(100), ncol = 2)
y <- rbinom(50, 1, 0.5)
fit <- logistic_fit(X, y)
pred <- predict(fit)

}
