% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_fit.R
\name{print.logistic_fit}
\alias{print.logistic_fit}
\title{Print method}
\usage{
\method{print}{logistic_fit}(x, ...)
}
\arguments{
\item{x}{A logistic_fit object}

\item{...}{Not used}
}
\description{
Print method
}
